PD#132697: OMX: fix 64bit cannot set pts

+++ tvp/OmxUtil.cpp
@@ -16,8 +16,8 @@
 
 #define AMSTREAM_IOC_MAGIC  'S'
 
-#define AMSTREAM_IOC_SET_OMX_VPTS  _IOW(AMSTREAM_IOC_MAGIC, 0xaf, unsigned long)
-#define AMSTREAM_IOC_SET_VIDEO_DISABLE  _IOW(AMSTREAM_IOC_MAGIC, 0x49, unsigned long)
+#define AMSTREAM_IOC_SET_OMX_VPTS  _IOW(AMSTREAM_IOC_MAGIC, 0xaf, int)
+#define AMSTREAM_IOC_SET_VIDEO_DISABLE  _IOW(AMSTREAM_IOC_MAGIC, 0x49, int)
 
 static int amvideo_handle = 0;
